fix: update regex pattern for e-mail inputs
The previous pattern showed that `user@email` is a valid e-mail address,
but the validator failed to pass, so this tweak will resolve that issue.

Which one is correct (in more cases) is not something I can tell with
certainty. The previous pattern was taken from W3C's documentation.

Here is a read about "dotless" domains from 2013, and why they're
prohibited: https://www.icann.org/news/announcement-2013-08-30-en