feat: reroute external updateURL fields to our API
In order to prevent various exploits that could happen with updateURL,
this is the simplest solution that I could come up with thus far.

While I'm not a huge fan of doing this, it is necessary for Stylus
integration, and it is reversible in case things need to change.

References openstyles/stylus#1223.