fix(images): add more data to error logs
There's an issue with images which I can't reproduce when images are
generated, yet database column is updated. This commit aims to provide
better logs in order to debug where the root cause comes from.

I have an idea why and where it happens, but I'll wait until there is
more data to work with.