fix(api): clone ID parameter in code endpoint
Attempting to work around an issue caused by Fiber's "zero allocation"
in go routines when upserting style statistics. If this commit works as
intended, there should be a slight increase across all style statistics.

Last but not least, huge thanks to @Gusted for writing the tests as well
as linking to the PR that implements the "UnsafeClone" function.