chore: switch fiber depedency
We're going to switch the `gofiber/fiber`[1]
dependency our main dependency in userstyles.world to have routing, simple usage of `fasthttp`etc. to our own fork of it.

Reason:
A few months back I've observed a interesting PR[2]
A PR that will help improve our security! 

However the contributor doesn't seem to work on the PR, and thus I've taken it on to fork fiber and apply the code into the main branch of `userstyles-world/fiber`[3] so we can enjoy this PR(with a few tweaks of the original code) and see the usage and implementation against real-world data and against attackers. And hopefully finally have it merged so I don't have to maintain a fork + patches.

1. https://github.com/gofiber/fiber
2. https://github.com/gofiber/fiber/pull/1247
3. https://github.com/userstyles-world/fiber