security(random): replace randomness with /dev/urandom
It was possible in a finite amount of time to brute force the correct seed(with the help of discord bot which happily shows our uptime) and generate all numbers in ahead. This attack isn't detected to be abused and would require a good dedicated time to actually exploit this. We're switching to `/dev/urandom` for now to generate secure numbers until a better approach is available which can resist such attack but still be performant as before on larger amounts of randomness.