feat(user): save user's last login date
This is going to be used for a feature that @decembre suggested a while
back.  Namely, the idea is to show change in stats over time on the
upcoming user dashboard since the last login date.

I don't plan on making this field public information.  The intent is
purely functional in nature, because there's no better to do it AFAIK.

References #65