refactor(web): unify files in a shared directory
Due to the nature of how `go:embed` works, I've decided to move all
files related to front-end and put them in a "web" directory.  I also
intend to do the same to files from docs directory, but I'll do so in a
separate commit.  The naming is kinda off, but that's trivial to change.