fix(images): remove artifacts from WebP images
Adding this flag seems to not only remove the artifacts, but also reduce
the image size even further — up to 40% in some cases.  I think that's
most likely because we're using an old version of libwebp.  In the near
future, after we switch to another OS, we'll have to revisit it.