feat(images): improve caching of preview images
We might want to create a struct in order to skip over passing 5
arguments to the function, 1–2 of which aren't even needed in some
cases.  Additionally, we might want to embrace the type system.

The mechanism for image caching is pretty simple and has a bunch of
benefits over hashing file names.  We might need to tweak and/or change
it in the future, but it doesn't seem necessary at the moment.

Fixes #113