feat(images): implement preview thumbnails
This will drastically cut down on the amount of bandwidth, as well as
improve the loading speed, at the cost of some image clarity.  I went
with one thumbnail image, at least for the time being, because the size
of Stylus' inline search images, as well as card thumbnails, is pretty
similar.  The difference is around 16x31px or so, since we use a 16:9
image format.

Also, we'll generate both WebP and JPEG variants for the time being.  If
we see there's no use for the latter, we could easily remove it.

Additionally, we might want to add PreviewThumb(nail) field to database
in order to reduce the amount of needless processing that needs to be
done.  Alas, that's a discussion for a more appropriate place.

Fixes #77