fix(styles): resolve incorrect/broken preview URLs
By doing this, we won't end up setting a broken preview URL in the
database, because that's the way that `images.Generate()` works.

Namely, since we return a nil if file or preview variables don't have
any non-nil values — which is an indicator that the function ran without
errors — it would end up setting an incorrect image URL.  I could've
added a new error and ignored it in checks, so that it doesn't get
logged, but it felt kinda hack-y.

Maybe I've over-engineering it, but this feels like an appropriate way
to go about implementing this functionality.  Alas, we'll probably go
over it at a later date.