refactor(tools): split different dependency checks
I'd like to make a use of `run build prod` in CI and this commit helps
with that.  Namely, we don't depend on two of our dependencies during
build-time, so I'm moving them to where they're actually being used.

While it's not a big deal to always check/install them, it's a good idea
to introduce a separation of concerns.  We use `air` as a development
dependency and `vips` as a runtime dependency.  The rest of them are
fine, because they're all build-time dependencies.