feat(storage): optimize search engine re-indexing
This commit also helped reduce memory usage from ~300 MB — when indexing
around 4000 userstyles — down to ~60 MB.  When Go's GC is happy, @Gusted
is happy, too.  No more brrr!  (=

In addition, we could play around with the amount of items that are
being processed during runs to find a sweet spot, but 250 seems plenty.
We might also want to add some sort of delay between runs in the event
it's the CPU that goes brrr in production.