feat(ui): switch to relative timestamps everywhere
I got tired of remembering time formats, all of which have different
downsides to them.  By switching to the relative format, it should be
more user-friendly and consistent throughout the UI.

The only thing left is to implement a way to convert absolute time from
`datetime` attributes to localized time, and show that using a tooltip
once the user hovers over the relative timestamp.