fix(search): resolve a bug with too many results
With this commit, we'll be returning proper results in case there are
more than 96 matches in our search index.  The results amount returned
by our search engine has been increased to a maximum of 500 and amount
of cards rendered on a page remains at 96 for the time being.

Next step would be to add pagination.  We probably have everything set
in place for it, someone just needs to wire it all together.

Thank you Michael for pointing out irregularities in search results.