refactor(stats): simplify upserting views/installs
Part one of two (or three) for optimizing this functionality.  There is
an impact on performance, but it's negligent at the moment—we don't get
enough requests on either endpoint to witness significant performance
degradation.

In fact, response times were slightly higher under continuous load all
while handling 10 times more traffic.  Of course, this will more than
likely vary in production because our VPS isn't that powerful.

Moreover, there are better ways to optimize things—coming in later
parts—than throwing goroutines at the problem and calling it a day.
This is possibly one of those premature optimizations that went wrong,
but it's hard to comment on it without getting "real-world" data first.