feat(cache): introduce caching for installs/views
This change should reduce a bunch of I/O-related issues since we're
running upsert queries non-stop and it's becoming increasingly more
difficult to keep up as we get more requests.

This change is especially effective during bursts when a given user has
a dozen or more userstyles installed because they get updated at once.
We could ask Stylus developers to introduce a short delay—which would be
practically unnoticeable to users—if this commit isn't enough.  I ran a
ton of tests and that shouldn't be necessary with this change, but it
would've been necessary with the old way.