fix(oauth): add workaround for duplicate usernames
This change adds a randomized four-digit number if there's an existing
username.  Since we have three OAuth providers, with more coming in the
near future, this will allow new users to be created.  We'll need to add
a way to rename usernames pretty soon.

This change also adds a fail in case when there's a duplicate email
address.  We'll need to figure out what to do about this.  As someone
suggested a while ago, perhaps we could send an email notice, letting
the user know what has happened.

Fixes #132