chore(db): add a migration to reset histories
This is required for my upcoming commit, otherwise there will be a bug
that results in incorrect stats for all userstyles created before today,
because we're missing snapshots of stats for over a year.  This commit
removes all snapshots except the last one while keeping necessary stats.

It's supposed to be temporary (until stats are rewritten), but let's not
forget the following saying: "Nothing is more permanent than a temporary
solution."

Ref #112