fix(db): improve performance when querying stats
This comes with a disadvantage of not having real-time stats.  It looks
like a worthwhile tradeoff, at least for the time being.  Perhaps we can
explore other options in the future, including the rewrite of stats that
will drastically simplify and separate concerns to separate tables.

What's more, this change will also speed up generation of compact index.
With styles that I have locally, it went from ~600ms down to 250–300ms.

Ref #112