fix(search): adjust weights for search fields
The goal here is to give more weight to titles and descriptions as a way
to return more relevant results.  I'm sure this is merely the beginning
of tweaking that value, especially considering that a lot of userstyles
are still using old, pre-validation descriptions, which are filled with
all sorts of irrelevant data.

I've also fixed excluding that default sorting method, which I initially
overlooked.  It's good enough for now, until we get around rewriting
pagination to include the search parameters.

Ref #230