fix(search): remove substring functionality
As it turns out, it caused more weirdness than I imagined.  Therefore,
let's see where things go without substring search.  If that doesn't
improve the status quo either, we'll go back to the drawing board.

To simulate substring search, you can use a prefix with an asterisk.
For example, looking up all Git forges can be done with `git*`, which
will return all userstyles that contain a word starting with "git".

However, using `-` still has the same behavior.  To work around it, wrap
it with quotes, like `"dark-"`, which will return all userstyles that
contain "dark-" literally.

Both of these can be used to target specific columns: name, description,
notes, categories (added in a commit from the future), as well as other
columns we'll add later on (if we're happy with this approach).

Fix #233