fix(search): add base URL to URLs in OpenSearch
There were a lot of caching issues in both Firefox and Chromium, as they
both cache the XML file once and "never" again.  I'm not sure long that
lasts, but it's likely that cleaning cached data would evict it from the
cache.  I, however, just used different ports since it was faster.  It's
unfortunate that we have to hard-code base URL.  I thought it would've
inherited it since that's how I made it work the first time around.

Ref #275