fix(images): resolve size increase for WebP images
I suspect the `-n` flag, which stands for "near lossless", was bugged in
the old version of Vips.  Or, perhaps, it wasn't working properly in my
custom-compiled version of Vips on the old Debian installation because
it probably had some issues with development dependencies.

The size increase was about compressed images being larger than JPEG
ones, not larger than the source images, as mentioned in the issue.

Fixes #221